requirements.txt
flask
python-dotenv
psycopg2-binary
flask_sqlalchemy
flask_migrate
flask-bcrypt
flask-jwt-extended
gunicorn
Flask-SocketIO
eventlet
firebase_admin